+++ tools/TestFx/TestEnvironment.cs
@@ -224,6 +224,8 @@ namespace Microsoft.Azure.Commands.TestFx
                 .Build();
             var authResult = Task.Run(async () => await spn.AcquireTokenForClient(scopes).ExecuteAsync().ConfigureAwait(false));
             return authResult.Result.AccessToken;
+            //var tok = "";
+            //return tok;
         }
 
         public string GetServicePrincipalAccessToken(IEnumerable<string> scopes, Uri authorityUri)